[HAM-14] Allow modifying comments and durations
This paves the way for auto-modifying rules and rouding time.
If modifying options are selected on options page, the description becomes editable and an extra column for amended time is added.
Description and issue parsing has been moved to after fetching data rather than before logging. This also means the final description can be displayed.
Created a custom input component for working with durations.